mem.c: Update mem.c

+++ src/mem.h
@@ -7,8 +7,8 @@
 
 struct mem_info
 {
-    int used;
-    int total;
+    long long used;
+    long long total;
     float percent;
 };
